requirements.txt
tensorflow-cpu==2.16.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
python-dotenv==1.0.1
